fix(lang-nodejs): get type of installed deps.
Get the type of dependencies from npm-list instead of package-lock.json due to some dependencies do not actually install in the local machine.
This fix should avoid the problem that the NpmWrapper cannot get SHA-1 from package-lock.json.
However, the NpmWrapper cannot get the information of bundledDependencies and optionalDependencies anymore.